Fixes upload tests (MH-102)
Since some views require advancedirective they are made forbidden when
it doesn't exist.
